Moderation update - 3.0.0

+++ modules/datamanagement.js
@@ -2,8 +2,10 @@
 exports.writeData = async (collectionName, data, overWriteConditions) => {
 
     const schema = require(`./schemas/${collectionName}-schema.js`)
-
-    await schema.deleteMany(overWriteConditions)
+    if(overWriteConditions != undefined){
+        await schema.deleteMany(overWriteConditions)
+    }
+    
     await new schema(data).save()
